Updated navigation for how many children, max free hours info
