+++ app/uk/gov/hmrc/childcarecalculatorfrontend/views/maxFreeHoursInfo.scala.html
@@ -27,7 +27,7 @@
 
     <p>@messages("max.free.hours.info.can.get.maximum.hours.paragraph")</p>
 
-    <a class="button" id="next-button" href="@routes.MaxFreeHoursInfoController.onPageLoad">
+    <a class="button" id="next-button" href="@routes.HowManyChildrenController.onSubmit">
         @messages("next.button")
     </a>
 }
